+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the "direnv" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+### Fixed
+- Show error when direnv executable is missing
+### Added
+- Make path to direnv executable configurable
+- Offer to install direnv or configure the executable path when it's not found
 
 ## [0.4.0]
 ### Fixed